Use a long long to store debug type flags. Fix a bug when it is impossible to log debug data in i386.

+++ inc/com/centreon/engine/configuration/applier/logging.hh
@@ -62,7 +62,7 @@ namespace                configuration {
 
       com::centreon::logging::file*
                          _debug;
-      unsigned long      _debug_level;
+      unsigned long long _debug_level;
       unsigned long      _debug_max_size;
       unsigned int       _debug_verbosity;
       com::centreon::logging::file*
